rename and reorganize files
