+++ app/javascript/components/applications/ApplicationEditPage.js
@@ -3,8 +3,8 @@ import React from 'react'
 import mapProps from 'utils/mapProps'
 
 import CardLayout from '../layouts/CardLayout'
-import { saveApplication } from './actions'
 import PaperApplicationForm from './application_form/PaperApplicationForm'
+import { saveApplication } from './applicationRequestUtils'
 
 const ApplicationEditPageForm = ({ listing, application, editPage, lendingInstitutions }) => {
   const saveEditApplication = async (submitType, submittedValues, listing, editPage) => {
