+++ app/javascript/components/applications/applicationRequestUtils.js
