+++ app/javascript/components/lease_ups/LeaseUpApplicationsTable.js
@@ -5,10 +5,10 @@ import { trim } from 'lodash'
 import ReactTable from 'react-table'
 
 import Button from 'components/atoms/Button'
+import { applicationsTablePageChanged } from 'components/lease_ups/actions/actionCreators'
 import CheckboxCell from 'components/lease_ups/application_page/CheckboxCell'
 import PreferenceRankCell from 'components/lease_ups/application_page/PreferenceRankCell'
 import StatusCell from 'components/lease_ups/application_page/StatusCell'
-import { applicationsTablePageChanged } from 'context/actionCreators/actionCreators'
 import { useAppContext } from 'utils/customHooks'
 import { MAX_SERVER_LIMIT } from 'utils/EagerPagination'
 import { cellFormat } from 'utils/reactTableUtils'
