+++ app/javascript/components/lease_ups/utils/shortFormRequestUtils.js
