+++ app/javascript/components/supplemental_application/utils/supplementalApplicationUtils.js
