+++ app/javascript/routes/LeaseUpRoutes.js
@@ -5,7 +5,7 @@ import { Switch, Route } from 'react-router-dom'
 import ApplicationPage from 'components/applications/ApplicationPage'
 import LeaseUpApplicationsPage from 'components/lease_ups/LeaseUpApplicationsPage'
 import LeaseUpListingsPage from 'components/lease_ups/LeaseUpListingsPage'
-import ApplicationDetailsPage from 'components/supplemental_application/ApplicationDetailsPage'
+import SupplementalApplicationPage from 'components/supplemental_application/SupplementalApplicationPage'
 import appPaths from 'utils/appPaths'
 
 // List of URLs that need to be routed via rails but that conflict with react-routed
@@ -29,7 +29,7 @@ const LeaseUpRoutes = () => (
       <LeaseUpListingsPage />
     </Route>
     <Route exact path={appPaths.toLeaseUpApplication(':applicationId')}>
-      <ApplicationDetailsPage />
+      <SupplementalApplicationPage />
     </Route>
     <Route exact path={appPaths.toApplication(':applicationId')}>
       <ApplicationPage />
