+++ spec/javascript/components/lease_ups/utils/leaseUpRequestUtils.test.js
@@ -2,7 +2,7 @@ import {
   convertToCommaSeparatedList,
   getApplications,
   sanitizeAndFormatSearch
-} from 'components/lease_ups/leaseUpActions'
+} from 'components/lease_ups/utils/leaseUpRequestUtils'
 
 const mockFetchLeaseUpApplications = jest.fn()
